requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
email-validator
pydantic
cryptography